ENH+TST: split Fraunhofer FKIE DDoS attack feed parser
