+++ docs/Feeds.md
@@ -914,6 +914,55 @@ To add feeds to this file add them to `intelmq/etc/feeds.yaml` and then run
 
 # Fraunhofer
 
+## DDoS Attack Feed (C&C)
+
+* **Status:** on
+* **Revision:** 01-07-2018
+* **Description:** The Fraunhofer DDoS attack feed provides information about tracked C&C servers and detected attack commands from these C&Cs.
+
+### Collector
+
+* **Module:** intelmq.bots.collectors.http.collector_http_stream
+* **Configuration Parameters:**
+*  * `http_password`: `{{ your password }}`
+*  * `http_url`: `https://feed.caad.fkie.fraunhofer.de/ddosattackfeed`
+*  * `http_username`: `{{ your username }}`
+*  * `name`: `DDoS Attack Feed (C&C)`
+*  * `provider`: `Fraunhofer`
+*  * `rate_limit`: `10`
+*  * `strip_lines`: `True`
+
+### Parser
+
+* **Module:** intelmq.bots.parsers.fraunhofer.parser_ddosattack_cnc
+* **Configuration Parameters:**
+*  * `unknown_messagetype_accuracy`: `80`
+
+
+## DDoS Attack Feed (Targets)
+
+* **Status:** on
+* **Revision:** 01-07-2018
+* **Description:** The Fraunhofer DDoS attack feed provides information about tracked C&C servers and detected attack commands from these C&Cs.
+
+### Collector
+
+* **Module:** intelmq.bots.collectors.http.collector_http_stream
+* **Configuration Parameters:**
+*  * `http_password`: `{{ your password }}`
+*  * `http_url`: `https://feed.caad.fkie.fraunhofer.de/ddosattackfeed`
+*  * `http_username`: `{{ your username }}`
+*  * `name`: `DDoS Attack Feed (Targets)`
+*  * `provider`: `Fraunhofer`
+*  * `rate_limit`: `10`
+*  * `strip_lines`: `True`
+
+### Parser
+
+* **Module:** intelmq.bots.parsers.fraunhofer.parser_ddosattack_target
+* **Configuration Parameters:**
+
+
 ## DGA Archive
 
 * **Status:** on
