+++ intelmq/bots/BOTS
@@ -417,12 +417,17 @@
             "module": "intelmq.bots.parsers.dyn.parser",
             "parameters": {}
         },
-        "Fraunhofer DDoS Attacks": {
-            "description": "Fraunhofer DDoS Attack Feed Parser is the bot responsible to parse the report and sanitize the information.",
-            "module": "intelmq.bots.parsers.fraunhofer.parser_ddosattack",
+        "Fraunhofer DDoS Attack C&C": {
+            "description": "Fraunhofer DDoS Attack C&C Feed Parser is the bot responsible to parse the report and sanitize the information.",
+            "module": "intelmq.bots.parsers.fraunhofer.parser_ddosattack_cnc",
             "parameters": {
                 "unknown_messagetype_accuracy": 10.0
             }
+        },
+        "Fraunhofer DDoS Attack Target": {
+            "description": "Fraunhofer DDoS Attack Target Feed Parser is the bot responsible to parse the report and sanitize the information.",
+            "module": "intelmq.bots.parsers.fraunhofer.parser_ddosattack_target",
+            "parameters": {}
         },
         "Fraunhofer DGA": {
             "description": "Fraunhofer DGA Parser is the bot responsible to parse the report and sanitize the information.",
